[Team-5] Dino, Bazzi 4주차 중간 PR

+++ backend/src/main/java/com/team5/issue_tracker/label/dto/response/LabelPageResponse.java
@@ -10,8 +10,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class LabelPageResponse {
-  private Long total;
-  private Long page;
-  private Long perPage;
+  private Integer total;
+  private Integer page;
+  private Integer perPage;
   private List<LabelResponse> labels;
 }
